fix(routing): init expo routing & merge screen into apps
