+++ app/(tabs)/_layout.js
@@ -1,0 +1,7 @@
+import { Stack } from 'expo-router';
+
+export default function TabLayout() {
+  return (
+    <Stack screenOptions={{ headerShown: false }} />
+  );
+}
