+++ app/_layout.js
@@ -1,0 +1,10 @@
+import { Stack } from "expo-router";
+import "../global.css";
+
+export default function RootLayout() {
+  return (
+    <Stack screenOptions={{ headerShown: false }}>
+      <Stack.Screen name="index" options={{ headerShown: false }} />
+    </Stack>
+  );
+}
