+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "mobile-rekbr-app",
   "version": "1.0.0",
-  "main": "index.js",
+  "main": "expo-router/entry",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
@@ -21,7 +21,7 @@
     "expo": "~53.0.11",
     "expo-clipboard": "^7.1.4",
     "expo-font": "~13.3.1",
-    "expo-router": "~5.0.7",
+    "expo-router": "~5.1.0",
     "expo-status-bar": "~2.2.3",
     "install": "^0.13.0",
     "lucide-react-native": "^0.514.0",
